check weather there is tables exists in db before implement foreign key constraints

+++ database/migrations/9999_12_30_000001_add_foreign_key_constraints_to_all_tables.php
@@ -13,13 +13,34 @@ class AddForeignKeyConstraintsToAllTables extends Migration
      */
     public function up()
     {
-        Schema::table('users', function (Blueprint $table) {
-            $table->foreign('designation_id')->references('id')->on('designations')->onDelete('cascade');
-        });
+        
+        //db tables belongs to app      - not check existance
+        //db tables belongs to modules  - check existance
+        
+        //users           - db tables belongs to app 
+        //courses         - db tables belongs to modules 
+        //projects,tasks  - db tables belongs to modules
 
-        Schema::table('courses', function (Blueprint $table) {
-            $table->foreign('author_id')->references('id')->on('users')->onDelete('cascade');
+        Schema::table('users', function (Blueprint $table) {
+            if (Schema::hasTable('designations')) {
+                $table->foreign('designation_id')->references('id')->on('designations')->onDelete('cascade');
+            }
         });
+        
+        if (Schema::hasTable('courses')) {
+            Schema::table('courses', function (Blueprint $table) {                
+                $table->foreign('author_id')->references('id')->on('users')->onDelete('cascade');
+            });
+        }       
+        
+        /*
+        if (Schema::hasTable('projects') && Schema::hasTable('tasks')) {
+            Schema::table('tasks', function (Blueprint $table) {                
+                $table->foreign('project_id')->references('id')->on('projects')->onDelete('cascade');
+            });
+        }
+        */
+
     }
 
     /**
@@ -29,13 +50,25 @@ class AddForeignKeyConstraintsToAllTables extends Migration
      */
     public function down()
     {
-        Schema::table('users', function (Blueprint $table) {
-            $table->dropForeign(['designation_id']);
+        Schema::table('users', function (Blueprint $table) {            
+            $table->dropForeign(['designation_id']);                      
         });
+        
+        if (Schema::hasTable('courses')) {
+            Schema::table('courses', function (Blueprint $table) {
+                $table->dropForeign(['author_id']);
+            });
+        }
+        
+        /*        
+        if (Schema::hasTable('tasks')) {
+            Schema::table('tasks', function (Blueprint $table) {         
+                $table->dropForeign(['project_id']);                      
+            });
+        }
+        */
 
-        Schema::table('courses', function (Blueprint $table) {
-            $table->dropForeign(['author_id']);
-        });
+
     }
 }
